removed redundancy by moving code to a function within Piece rather than individual pieces
