+++ src/engine/pieces/piece.js
@@ -1,3 +1,5 @@
+import Square from "../square";
+
 export default class Piece {
     constructor(player) {
         this.player = player;
@@ -10,5 +12,26 @@ export default class Piece {
     moveTo(board, newSquare) {
         const currentSquare = board.findPiece(this);
         board.movePiece(currentSquare, newSquare);
+    }
+
+    loopOverRowsAndCols(board, rowsAndCols) {
+        // given 'this' [[rowDir1, colDir1], ...], loop through all available moves
+        let position = board.findPiece(this);
+        let available = [];
+        for (let [rowDir, colDir] of rowsAndCols) {
+            let row = position.row + rowDir;
+            let col = position.col + colDir;
+            let square = Square.at(row, col);
+            while (Square.isValid(row, col) && board.getPiece(square) === undefined) {
+                available.push(square);
+                row += rowDir;
+                col += colDir;
+                square = Square.at(row, col);
+            }
+            if (Square.isValid(row, col) && board.getPiece(square) && board.getPiece(square).player !== this.player) {
+                available.push(square);
+            }
+        }
+        return available;
     }
 }
